+++ .teamcity/_Self/buildTypes/GithubTests.kt
@@ -12,7 +12,6 @@ import jetbrains.buildServer.configs.kotlin.v2019_2.triggers.VcsTrigger
 import jetbrains.buildServer.configs.kotlin.v2019_2.triggers.vcs
 
 object GithubTests : Github("clean test", "Tests")
-object GithubLint : Github("clean ktlintCheck", "Lint")
 
 sealed class Github(command: String, desc: String) : BuildType({
   name = "GitHub Pull Requests $desc"