+++ src/com/maddyhome/idea/vim/action/motion/leftright/TillCharacterMotion.kt
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
+
 package com.maddyhome.idea.vim.action.motion.leftright
 
 import com.intellij.openapi.actionSystem.DataContext
@@ -24,16 +25,37 @@ import com.maddyhome.idea.vim.VimPlugin
 import com.maddyhome.idea.vim.command.Argument
 import com.maddyhome.idea.vim.command.CommandFlags
 import com.maddyhome.idea.vim.command.MotionType
-import com.maddyhome.idea.vim.group.MotionGroup
 import com.maddyhome.idea.vim.handler.Motion
 import com.maddyhome.idea.vim.handler.MotionActionHandler
+import com.maddyhome.idea.vim.handler.toMotionOrError
+import com.maddyhome.idea.vim.helper.Direction
 import com.maddyhome.idea.vim.helper.enumSetOf
 import java.util.*
 
-class MotionLeftMatchCharAction : MotionActionHandler.ForEachCaret() {
+enum class TillCharacterMotionType {
+  LAST_F,
+  LAST_SMALL_F,
+  LAST_T,
+  LAST_SMALL_T,
+}
+
+class MotionLeftMatchCharAction : TillCharacterMotion(Direction.BACKWARDS, TillCharacterMotionType.LAST_F, false)
+class MotionLeftTillMatchCharAction : TillCharacterMotion(Direction.BACKWARDS, TillCharacterMotionType.LAST_T, true)
+class MotionRightMatchCharAction : TillCharacterMotion(Direction.FORWARDS, TillCharacterMotionType.LAST_SMALL_F, false)
+class MotionRightTillMatchCharAction :
+  TillCharacterMotion(Direction.FORWARDS, TillCharacterMotionType.LAST_SMALL_T, true)
+
+sealed class TillCharacterMotion(
+  private val direction: Direction,
+  private val tillCharacterMotionType: TillCharacterMotionType,
+  private val finishBeforeCharacter: Boolean,
+) : MotionActionHandler.ForEachCaret() {
   override val argumentType: Argument.Type = Argument.Type.DIGRAPH
 
   override val flags: EnumSet<CommandFlags> = enumSetOf(CommandFlags.FLAG_ALLOW_DIGRAPH)
+
+  override val motionType: MotionType =
+    if (direction == Direction.BACKWARDS) MotionType.EXCLUSIVE else MotionType.INCLUSIVE
 
   override fun getOffset(
     editor: Editor,
@@ -43,13 +65,14 @@ class MotionLeftMatchCharAction : MotionActionHandler.ForEachCaret() {
     rawCount: Int,
     argument: Argument?,
   ): Motion {
-    if (argument == null) {
-      return Motion.Error
+    if (argument == null) return Motion.Error
+    val res = if (finishBeforeCharacter) {
+      VimPlugin.getMotion()
+        .moveCaretToBeforeNextCharacterOnLine(editor, caret, direction.toInt() * count, argument.character)
+    } else {
+      VimPlugin.getMotion().moveCaretToNextCharacterOnLine(editor, caret, direction.toInt() * count, argument.character)
     }
-    val res = VimPlugin.getMotion().moveCaretToNextCharacterOnLine(editor, caret, -count, argument.character)
-    VimPlugin.getMotion().setLastFTCmd(MotionGroup.LAST_F, argument.character)
-    return if (res < 0) Motion.Error else Motion.AbsoluteOffset(res)
+    VimPlugin.getMotion().setLastFTCmd(tillCharacterMotionType, argument.character)
+    return res.toMotionOrError()
   }
-
-  override val motionType: MotionType = MotionType.EXCLUSIVE
 }