+++ src/com/maddyhome/idea/vim/helper/CommandStateExtensions.kt
@@ -29,10 +29,23 @@ import com.maddyhome.idea.vim.option.OptionsManager
 val usesVirtualSpace
   get() = OptionsManager.virtualedit.value == "onemore"
 
+/**
+ * Please use `isEndAllowed` based on `Editor` (another extension function)
+ * It takes "single command" into account.
+ */
 val CommandState.Mode.isEndAllowed: Boolean
   get() = when (this) {
     CommandState.Mode.INSERT, CommandState.Mode.VISUAL, CommandState.Mode.SELECT -> true
     CommandState.Mode.COMMAND, CommandState.Mode.CMD_LINE, CommandState.Mode.REPLACE, CommandState.Mode.OP_PENDING -> usesVirtualSpace
+  }
+
+val Editor.isEndAllowed: Boolean
+  get() = when (this.mode) {
+    CommandState.Mode.INSERT, CommandState.Mode.VISUAL, CommandState.Mode.SELECT -> true
+    CommandState.Mode.COMMAND, CommandState.Mode.CMD_LINE, CommandState.Mode.REPLACE, CommandState.Mode.OP_PENDING -> {
+      // One day we'll use a proper insert_normal mode
+      if (this.subMode == CommandState.SubMode.SINGLE_COMMAND) true else usesVirtualSpace
+    }
   }
 
 val CommandState.Mode.isEndAllowedIgnoringOnemore: Boolean